spec: define regular and alias type declarations
(This was formerly https://go-review.googlesource.com/#/c/35812/)

Now that a type declaration can declare both a new named type and
alternatively, a type alias for a given type, the notion of
a named type is becoming problematic. Given a name for a type, it
may or may not be a "named type" in the original sense; yet clearly
the type has a name and thus can be considered "named".

Specific changes:

- Introduced the notion of a "branded type": This is what formerly
  was called a "named type".

- Adjusted language that implicitly assumed a named type is different
  from an unnamed type: A type with a name may now be identical to
  a type literal if the type name is a type alias denoting that literal.
  Furthermore, two differently named types may denote the same branded
  type.

- Adjusted type declarations: There are now two forms, one of which
  creates a branded type first before the type name is bound to it.

Type declarations are now more in line with all the other declarations:
It's primary purpose is to bind a name to a type. In the specific (old)
form, it also creates a new branded (formerly) named type before binding
the name.

Change-Id: I54e3c488c90e6880c8bb55e9a13d8dc985914d4e