syscall: improve handling of netlink alignment
The existing implementation reports an error for messages whose size is
not a multiple of the netlink alignment and doesn't return them to the caller.
Per the netlink specification, messages are required to start at an aligned
offset, but are not required to end on it, leaving the possibility for the
last message to have an unaligned size.
This change allows the last message to have an unaligned size and documents  intent.

Fixes #18714

Change-Id: Iea17896024896d9a30ff478ecc7937c85af7fb01