cmd/compile/internal: use RBIT to simplify CTZ on ARMv7
I make two changes:

1. Introduce an "extraPkg" field in the struct arch in ssa/gen/main.go,
so the ssa/gen/rulgen.go can import different packages for each arch's
rewriteXXX.go. For example, import cmd/internal/obj for rewriteARM.go.

2. By checking GOARM in ssa/gen/ARM.rules, each intermediate operator
can be implemented via different instructions. For example, CTZ is
simply implemented with RBIT-CLZ on ARMv7, otherwise it
becomes more complex with a serial of ARMv6 instructions.

It is up to the user to choose compitability or efficiency.

Change-Id: Ie9ee645fa39333fa79ad84ed4d1cefac30422814