spec: define alias type declarations, notion of "original" type
(This was formerly https://go-review.googlesource.com/#/c/35843/)

This is an alternative to https://go-review.googlesource.com/36020:
The organization is the same but instead of using the term "branded" type,
it uses the notion of an "original" named type (vs other named types). An
original named type (or just original type), corresponds to what we used
to (originally) call a named type; but "original" also reflects that fact
that these types are newly created, and different from all other types.

Change-Id: Ifc774de065adcce41ecb4eef2fac7e3e03956251