Update patch set 3
Patch Set 3:

> I agree there's a problem here, but I wonder if we can somehow
 > avoid referring to the alias as a name instead. "Branded" doesn't
 > seem right, and I looked through various synonym lists without
 > finding anything good either.
 > 
 > For example, we could define that
 > a type can be denoted in three different ways: by a literal, by a
 > name (introduced by type T1 type-expr), and by an alias (introduced
 > by type T1 = type-expr).
 > 
 > This separates that there are names and there are aliases. That's
 > actually a distinction in real life too, so it might be sustainable
 > here.

Ok, I'll set up a 3rd CL trying your suggestion and see how that turns out. An issue I anticipate is how to deal with the fact that an alias may actually denote a name. Another one is that both aliases and names are visually names. I've tried to make that very clear by separating the concept of a name from the concept of a type, which is actually what's happening in the implementation. The implementation doesn't have alias names, it just has names. But again, I'll give it shot.

If one of the concerns is the term "branded", please consider https://go-review.googlesource.com/#/c/35843/, which is using "original type" instead (which matches some of terminology we used for the 1.8 alias implementation).

Patch-set: 3