crypto/tls: document ConnectionState.NegotiatedProtocol more clearly
NegotiatedProtocol's documentation implies that the protocol will always
be from Config.NextProtos but with server side NPN, a client can select
a protocol that the server does not support. This commit clarifies that
there is guarantee.

I also added a note to the documentation for
ConnectionState.NegotiatedProtocolIsMutual, making it clear that it is
client side only.

Change-Id: Icd028af8042f31e45575f1080c5e9bd3012e03d7