crypto/tls: document ConnectionState.NegotiatedProtocol more clearly
This bit me hard, anyone could crash my server if they used NPN instead
of ALPN. I was under the assumption that the NegotiatedProtocol could
only come from Config.NextProtos but because of NPN, there is no
guarantee.

It was unclear because how can a protocol that the server did not
advertise be considered a negotiated? This makes it clear.

I also added a note to the documentation for
ConnectionState.NegotiatedProtocolIsMutual, making it clear that it is
client side only.

Change-Id: Icd028af8042f31e45575f1080c5e9bd3012e03d7