spec: introduce type alias declarations
This is yet another attempt to introduce type alias declarations
in the spec. Instead of introducing new terminology explicitly,
the notion of a "named type" was removed. Instead, the spec refers
to the kind of declaration introducing a type where a distinction
is needed.

Still, to be able to talk about what used to be called "named type",
some terminology is needed. At the moment this is "new, distinct" type.

Change-Id: I1f8889e42fc895834a2b37d3f98f2870fa2960e2