Update patch set 2
Patch Set 2:

(1 comment)

I still think Modula-3's "branded" is the best term we've seen so far to describe a concept that was previously bound up with namedness but, with the addition of aliases, is no longer.  It is unfortunate that adding a concept to the language requires not just the addition of a new term for that concept, but the renaming of existing concepts.  That said, the future is longer than the past, and those who will learn from the Go spec outnumber those who have already done so.

Patch-set: 2
Label: Code-Review=0