cmd/compile: check loop rescheduling with stack bound, not counter
After benchmarking with a compiler modified to have better
spill location, it became clear that this method of checking
was actually faster on (at least) two different architectures
(ppc64 and amd64) and it also provides more timely interruption
of loops.

This change adds a modified FOR loop "FORUNTIL" that checks
after executing the loop body instead of before (i.e., always
at least once).  This ensures that a pointer past the end
of a slice or array is not made visible to the garbage collector.

Change-Id: Ieafaa9a61d2a583ad00968110ef3e7a441abca50