cmd/compile: check loop rescheduling with stack bound, not counter
After benchmarking with a compiler modified to have better
spill location, it became clear that this method of checking
was actually faster on (at least) two different architectures
(ppc64 and amd64) and it also provides more timely interruption
of loops.

This change adds a modified FOR loop node "FORUNTIL" that
checks after executing the loop body instead of before (i.e.,
always at least once).  This ensures that a pointer past the
end of a slice or array is not made visible to the garbage
collector.

This change apparently provides a 1% geomean improvement on
PPC64 running the go1 benchmarks; the improvement on AMD64
is only 0.12%.  These are improvements observed without the
rescheduling experiment; the restructured loop has had an
unconditional branch removed from each iteration.

Change-Id: Ieafaa9a61d2a583ad00968110ef3e7a441abca50