cmd/compile: fix error message for missing brace in if statement
Fixes #18747.

Fixes up with a nicer error message when an if statement
has a missing brace. There are 3 cases to be considered:
1.
if {
}
or
if
{
}
or
if token; {
}
In this case tell them that there is a missing condition
if the if statement, as we have always done

2.
if condition
without the brace on the same line as the condition.
In this case error out with "missing { after if clause".

3.
if ()
Unexpected ), expecting expression.
However case 3) is blocked by https://golang.org/issues/18915
since it gives a spurious/unexpected error message, so that test
case hasn't been included in this bug test.

Case 2) is handled by this CL. I found that I couldn't get access
to the previous expression, since it was consumed in the if statement
line parsing. Therefore the hack to disambiguate case 1) from 2)
was that in case 1) they are on the same line.

Change-Id: I647e77075b3387b2c1cdb9999ec5759bd812b58b