Update patch set 1
Patch Set 1: -Run-TryBot

I looked at the objdump and saw that there 93 not,and pairs replaced by andc in the keccafF1600 function.  There was an improvement of 5-7% in the sha3 benchmarks.  If you want me to post real results I'd have to redo.  

This occurs in other places too but sha3 was reported to me first.

Patch-set: 1
Label: Run-TryBot=0