runtime: use atomic operations for fwdSig
The fwdSig array is accessed by the signal handler, which may run in
parallel with other threads manipulating it via the os/signal package.
Use atomic accesses to ensure that there are no problems.

Change the signal handler to not access the sigtab array, as it
doesn't need to. It was checking whether _SigHandling was set, but if
_SigHandling is clear the signal handler would not be installed and
would not run anyhow.

Change-Id: Icfa407518c4ebe1da38580920ced764898dfc9ad