runtime: use atomic ops for fwdSig, make sigtable immutable
The fwdSig array is accessed by the signal handler, which may run in
parallel with other threads manipulating it via the os/signal package.
Use atomic accesses to ensure that there are no problems.

Move the _SigHandling flag out of the sigtable array. This makes sigtable
immutable and safe to read from the signal handler.

Change the signal handler to not check whether the signal is being
handled. If it were not being handled, the signal handler would not be
run anyhow.

Change-Id: Icfa407518c4ebe1da38580920ced764898dfc9ad