Update patch set 1
Patch Set 1:

This is the first patch I've made for the runtime so I apologize for any major mistakes. I made sure all Go tests passed with -race as well as my own trace decode/encoder package that has fairly extensive test coverage. If the proposed trace event changes are accepted I would be glad to modify the patch to fix any issues, that said I also won't be offended if it's not even a viable starting point :- )

Patch-set: 1