Update patch set 2
Patch Set 2: Code-Review-2

(2 comments)

I'm not in favor of this. Yes, the API could be better, but in many of the use cases we actually have (compiler, go/constant), the current API is fine. There's usually one place where we need to make comparisons before we know if the value is in the desired range. It's not even a given that it has to be in the int64/uin64 range; sometimes we want smaller ranges and then we'd have to do another (admittedly cheaper) range check again.

I agree it would be useful to have these functions, but I'm not convinced that's a significant enough bar at this point.

Patch-set: 2
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2