Update patch set 3
Patch Set 3:

> FWIW, it still doesn't work for time.go. If you disasm Time.Equal, no BTQ is generated. The code there is "if value & 1<<63 {" which is supposed to be handled, as far as I can tell.

There are several things blocking that from working. First is a generic optimization:

// Rewrite AND of consts as shifts if possible, slightly faster for 64 bit operands
// trailing zeros can be shifted right, then left
(And64 <t> (Const64 [y]) x) && nlo(y) + ntz(y) == 64 && ntz(y) >= 32
  -> (Lsh64x64 (Rsh64Ux64 <t> x (Const64 <t> [ntz(y)])) (Const64 <t> [ntz(y)]))

That could be prevented for this case:

(And64 <t> (Const64 [y]) x) && nlo(y) + ntz(y) == 64 && ntz(y) >= 32 && (nlo(y) > 1 || !archHasBitTest(config))
  -> (Lsh64x64 (Rsh64Ux64 <t> x (Const64 <t> [ntz(y)])) (Const64 <t> [ntz(y)]))

The next obstacle is that our isPowerOfTwo doesn't recognize uint64(1)<<63 as a power of two, since it deals in int64s. That could also be fixed.

But then the pattern match also fails, because the &s don't get associated correctly, and it doesn't see that we are ANDing with a constant. We could add re-association/canonicalization rules to fix that, I guess.

But at this point I'm not sure it's worth the time+effort+risk to fix all those. If a use case comes along where this code or code like it is a bottleneck, I'll happily take another crack. (Or someone else is welcome to.)

Patch-set: 3