Update patch set 2
Patch Set 2:

I did do a lot of performance testing- I have another library I wrote that does trace encoding/decoding with a interface that wraps StartTrace and streams to a decoder. The trace throughput remained more or less the same over all, the differences from run-to-run vary a decent amount in general for anything that creates contention so it's hard to measure, but I can at least say there was no glaring performance degradation.

Probably because the locking exists before the changes as well, with the increase here being at most 2* the number of stacks. As the trace duration increases the cost of locks for stack events are diminished as more PCs are cached, which propagates to strings within frames as well.

A final note is that the data structure (Dmitry wrote it, I just generalized it to share it for string use) allows for concurrent reads on find. This really saves performance because it allows double-checked locking for cache lookups.

Patch-set: 2