Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > FWIW, it still doesn't work for time.go. If you disasm Time.Equal, no BTQ is generated. The code there is "if value & 1<<63 {" which is supposed to be handled, as far as I can tell.
> 
> There are several things blocking that from working. First is a generic optimization:
> 
> // Rewrite AND of consts as shifts if possible, slightly faster for 64 bit operands
> // trailing zeros can be shifted right, then left
> (And64 <t> (Const64 [y]) x) && nlo(y) + ntz(y) == 64 && ntz(y) >= 32
>   -> (Lsh64x64 (Rsh64Ux64 <t> x (Const64 <t> [ntz(y)])) (Const64 <t> [ntz(y)]))
> 
> That could be prevented for this case:
> 
> (And64 <t> (Const64 [y]) x) && nlo(y) + ntz(y) == 64 && ntz(y) >= 32 && (nlo(y) > 1 || !archHasBitTest(config))
>   -> (Lsh64x64 (Rsh64Ux64 <t> x (Const64 <t> [ntz(y)])) (Const64 <t> [ntz(y)]))
> 
> The next obstacle is that our isPowerOfTwo doesn't recognize uint64(1)<<63 as a power of two, since it deals in int64s. That could also be fixed.
> 
> But then the pattern match also fails, because the &s don't get associated correctly, and it doesn't see that we are ANDing with a constant. We could add re-association/canonicalization rules to fix that, I guess.

This part about re-association and canonicalisation seems like a general issue with rules. It's not the first time it comes up. Maybe it's worth discussing some general solution, like a canonical way of writing it (enforced during compilation?) and then a canon pass?

> But at this point I'm not sure it's worth the time+effort+risk to fix all those. If a use case comes along where this code or code like it is a bottleneck, I'll happily take another crack. (Or someone else is welcome to.)

To be frank, I consider this pattern matching a net win. The fact that it's hard to compose over other similar rules that have probably the same "small" effect but they were just committed before, sounds like a scalability problem of the rules framework, rather than just pinpointing a specific singularity that is not worth it.

For instance, the rule converting AND to SHL+SHR is really a minor gain and I doubt there was ever a benchmarking highlighting it as a hotspot. It was just committed earlier on, and happened not to conflict too much with other rules.

IOW I understand the principle of "too complex -> not worth" when you have a good design and architecture and you need to over-complicate it for a small benefit. But in this case, we just have hundreds of rules, and adding the Nth+1 conflicts in non trivial way.

As for the benchmark, Bloom filters relies on this kind of bit checking in the inner loop, when checking for the existence of an item. In fact, I found out about this BTQ optimisation on LLVM while reviewing a performance-optimised implementation of Bloom filters in C. The code in particular was implementing the so-called Scalable Bloom Filters, which is a chain of Bloom filters; to check for the existence of an item, the code needs to check for multiple bits in multiple filters in sequence, so the BTQ is really at the core of the algorithm.

Patch-set: 3