Update patch set 2
Patch Set 2:

This code doesn't make sense to me. What is going on? 
Why is there an atomic closed flag instead of a lock?
Why is there another atomic error value in addition to the atomic closed flag?
Multiple atomics is almost always a synchronization problem.
What are the constraints this code is attempting to solve?

This code is so complex it is not obviously wrong.
I would much rather it be obviously correct.

Patch-set: 2