Update patch set 1
Patch Set 1: Code-Review+2

As we discussed, let's merge this CL, and then you can rebase and merge the other CL on top of this one. For now we'll cherry-pick just this one into Go 1.8, and we've got the other one for a point release if we need it.

Patch-set: 1
Label: Code-Review=+2