cmd/pprof: separate discontiguous assembly blocks
Currently, weblist's disassembled output for each line simply lists
all instructions that are marked with that line number. This leads to
confusing disassembly because a lot of things look like straight-line
control flow when they aren't. For example, index panics look like
always happen:

   40ms       40ms   4206a3: LEAQ 0(R9)(R13*1), R14
   30ms       30ms   4206a7: CMPQ R10, R14
      .          .   4206aa: JAE 0x4208ad
      .          .   4206b0: MOVZX 0(BX)(R14*1), R15
   40ms       40ms   4206b5: MOVB R15, 0x47(SP)
      .          .   4208ad: CALL runtime.panicindex(SB)
      .          .   4208b2: UD2

In reality, 4208ad is at the end of the function, but it looks like we
call it immediately after a successful index operation.

Fix this by adding a vertical ellipses to separate blocks of assembly
that have other instructions between them. With this change, the above
looks like:

   40ms       40ms   4206a3: LEAQ 0(R9)(R13*1), R14
   30ms       30ms   4206a7: CMPQ R10, R14
      .          .   4206aa: JAE 0x4208ad
      .          .   4206b0: MOVZX 0(BX)(R14*1), R15
   40ms       40ms   4206b5: MOVB R15, 0x47(SP)
                          ⋮
      .          .   4208ad: CALL runtime.panicindex(SB)
      .          .   4208b2: UD2

Change-Id: I9ec446492c583f7fdb6d96d2251b350e695695a6