Update patch set 5
Patch Set 5:

BTW, I think the "if m < 0 {break" is unnecessary, because in any situation if "s" does not contain "sep", the loop just breaks (guarded by i<n). Particularly, the last part of "s" (no "sep") does not go into the loop. 
I tested this and it works for all cases.

Patch-set: 5