Update patch set 3
Patch Set 3:

Finally got nacl set up. All tests pass locally. Maybe it's the nacl version?

Brad, if you're reading this, do you know what version of nacl the builders have?


> This part about re-association and canonicalisation seems like a general issue with rules. It's not the first time it comes up. Maybe it's worth discussing some general solution, like a canonical way of writing it (enforced during compilation?) and then a canon pass?

There's been lots of discussion of commutative ops, and that has been solved somewhat generally. This is the first time I've seen association be an issue. That's not to say we shouldn't solve it, just that it's not the highest priority.


> > But at this point I'm not sure it's worth the time+effort+risk to fix all those. If a use case comes along where this code or code like it is a bottleneck, I'll happily take another crack. (Or someone else is welcome to.)
> 
> To be frank, I consider this pattern matching a net win. The fact that it's hard to compose over other similar rules that have probably the same "small" effect but they were just committed before, sounds like a scalability problem of the rules framework, rather than just pinpointing a specific singularity that is not worth it.

Yes it would be a net win. Yes, there are some rule scalability problems, perhaps most notably about the million ways to address things on amd64 (e.g. indexed offsets). But I don't have a good plan for addressing that.

And as for this particular rule, I'm not personally going to invest any more of my limited time on it now. I think I can have a bigger impact elsewhere. But I'd happily review CLs from you or anyone else if they want to tackle it.

Patch-set: 3