Update patch set 1
Abandoned

I believe that the appealing StdCmd delta reported in the commit message here is an artifact of using toolstash with compilebench. Any changes to this very expensive bit of code should probably be deeper algorithmic changes, not just specialization and removal of a few copies. Happy to restore and pursue further if there is disagreement.

Patch-set: 1
Status: abandoned
Tag: autogenerated:gerrit:abandon