Update patch set 3
Patch Set 3:

Managed to get this to fail locally. The problem is just that nacl doesn't know about the BT instruction, so it fails validation anywhere that the BT instruction gets used. So just skipping nacl on these rules is the right thing to do. This is ready to review.

Patch-set: 3