Update patch set 2
Patch Set 2:

You make a lot of good points Hyang-Ah, since it seems you guys are okay with the change itself (if implemented properly) I think the next step is to make changes to tracing measurable, would you agree?

I used a method similar to what you described but without the stress tool to generate large traces to make sure the go tool still worked, go test with -trace on the entire runtime produces a nice 1.2gb trace file for testing. Is there any places in the Go standard library that perform heavy stress tests that I could add something for tracing to?

Once we have a way to test and measure I think there may be some opportunity to implement this in a better way, it sort of evolved while I learned how everything worked and I think a fresh stab at it might be good. I think rather than changing the already correct and vetted logic for emitting events to write to the current buffer on demand, instead stage the strings/stacks to a separate buffer. This buffer could be linked to the head of the current batch and emitted before each batch is sent. I'm open for any other suggestions, but seeing that windows test fail makes me certain there is issues here.

Patch-set: 2