Update patch set 3
Patch Set 3:

I used ncval. It said there was an unrecognized instruction, and when I looked up the instruction at the offset specified, it was a BTQ (which was recognized as such by the hopper disassembler, so it was correctly encoded). From which I concluded that NCAA doesn't know about the bt instruction.

Patch-set: 3