sync: add Map and update several packages to use it
DO NOT REVIEW

TODO: finish review of x/sync version;
compare benchmarks of library changes

----

There are a few additional uses of RWMutex which could potentially be
replaced but would involve subtle or invasive changes:

syscall/fd_nacl.go uses an RWMutex to guard a map of reference-counted
files.  Use of sync.Map would require care to avoid races with
refcount updates.

syscal/env_unix.go uses an RWMutex to guard a map[string]int which
indexes into a []string.  sync.Map could potentially be used to reduce
contention in Getenv, but it would require either a parallel map
dedicated to that purpose or a significant change to the existing data
structures.

net/interface.go uses an RWMutex to guard a bidirectional map with
frequent writes.  It appears to acquire a read-lock along every path,
so it isn't obvious why it's using RWMutex in the first place.

text/template/template.go uses an RWMutex to guard a pair of maps
which are passed to parse.Parse.  Changing the underlying data
structure (without introducing needless copying on every Parse) would
require changes to the parse package API.

go/token/position.go uses an RWMutex to guard three variables with
fairly subtle invariants among them.  It is plausible that the RWMutex
could be replaced with a Map and one or more atomic variables, but the
changes to do so are not readily apparent.

net/dnsclient_unix.go uses an RWMutex to guard a single pointer.  It
should use an atomic variable instead.

crypto/tls/common.go uses an RWMutex to guard a pointer and a slice
which are each updated atomically.  It should use atomic variables
instead.

Change-Id: I63fa5ef6951b1edd39f84927d1181a4df9b15385