Update patch set 3
Patch Set 3:

> > And as for this particular rule, I'm not personally going to invest any more of my limited time on it now. I think I can have a bigger impact elsewhere. But I'd happily review CLs from you or anyone else if they want to tackle it.
> 
> Thanks. I wish I were up to speed with the framework, but it looks like it's hard for you, so I'm not sure how many hours it would take it to me :)

Like with any framework, some things are easy and others hard. :) And this isn't hard per se, it is mostly time-consuming.


> How do you find out which rules matched (like, how you did discover the rules in generic that matched)? And how do you debug why a rule doesn't match?

I mailed CL 36646, which is a tool I started hacking on specifically to figure out these match failures.

As for finding the rules in generic that matched, what I generally do when puzzling over these things is a combination of:

* find a minimal function
* compile with GOSSAFUNC=<functionname> and look at the ssa.html
* stare hard at what the opt and lower passes did
* grep the rules files for potentially relevant rules and run them in my head
* eliminate those rules and see what happens

So basically, old-fashioned debugging.

I have an idea for a better tool. I'll file an issue and cc everyone on this CL.

Patch-set: 3