Update patch set 3
Patch Set 3:

The new link is great. The bullet-list-sentence is not. That puts too much emphasis on that paragraph and is bad typography. The list in the paragraph seems perfectly readable, especially since it is just clarifying what "home directory" means, and most people already know. If you're worried about people not being able to skip over the list easily enough, please insert a paragraph break before the next sentence instead.

Patch-set: 3