cmd/compile: atomics as intrinsics on ppc64x
DO NOT SUBMIT

This adds the necessary changes so that atomics are treated as
intrinsics on ppc64x.  This also includes a change to the file
inl.go which is necessary to avoid a hang during the build,
but won't be in the final CL.

I'm creating this for now so the hang can be debugged.

Fixes #8739

Change-Id: Icb85e2755a49166ee3652668279f6ed5ebbca901