Update patch set 3
Patch Set 3:

Hello, I'm curious the requirements behind the version changes, I was thinking it was when format changes are made but it seems it may be anytime that events are added? Maybe it would be beneficial to consider a minor version to indicate the event count or a special initialization event to declare a new event type (or emit a count) that the parser will use to ignore events within this boundary. It doesn't seem to happen often so maybe not worth thinking about.

Patch-set: 3