Update patch set 2
Patch Set 2:

> > trace tests should go to src/runtime/trace dir.
 > 
 > Hmm I'm not sure how I could write a standard `go test` for a test
 > or benchmark, since they will take a good amount of time to run. So
 > my immediate thought was to write a main that executed tests with
 > the stress utility and produced a format compatible with the
 > benchcmp tool. Do you have an example where a similar problem like
 > this was solved that I could copy the pattern for? I can't find
 > one, sorry.

I don't know what you want to write, but generally it applies to all tests -- e.g. one may want sync.Mutex tests to run for hours. But it does not mean that everybody need to run them for hours all the time.
But add a test. Make it run for at most 0.1s in -short mode, make it run an order of magnitude more in non-short mode. Then use -count=1000 when you want to really stress it. But it steal needs to run on bots continuously, even if briefly.

Patch-set: 2
Label: Code-Review=0